+++ DDL/Tables/createTableSeat.sql
@@ -6,8 +6,8 @@ create table seat(
 	seat_id        int   not null auto_increment,
 	`number`       int   not null,
     airplane_id    int   not null,
-    class_id       int,
-    status_id      int   not null,
+    class_id       tinyint,
+    status_id      tinyint   not null,
     price          int,
    	
     constraint pk_seat_seat_id primary key (seat_id)
